requirements.txt
easydict
g2pk
onnx
onnxruntime-1.10.0
umap-learn
visdom
webrtcvad
numba
matplotlib
scipy
sounddevice
Unidecode
inflect
multiprocess
bokeh
jamo
pydub
kaldiio
ipykernel
starlette
uvicorn
jpype1
konlpy
nvsmi
jamotools
pygsheets
tweepy==3.10.0
nltk
pillow
jupyterlab
ptyprocess
pyzmq
lws
setuptools==59.5.0
anyascii
coqpit
gruut
jieba
pypinyin
fsspec
pyworld
librosa
wandb
tensorboardX
tensorboard
soundfile
numpy
natsort
tqdm
phonemizer
typeguard
g2p_en
pandas
openpyxl
nvidia-pyindex
sox
streamlit
starlette==0.17.1
gradio
